Updated Terms and Conditions
Changed the main blue 'Important notice' stating that we are FCA Approved when we are not. 

Also updated the fee from 3.5% to a minimum of 9%. This is inline with our needed corrections. 

Also make sure there was no <div left after the important blue shaded title stating relations with the FCA