Update Privacy
Mainly focused on getting the words changed from blue to white and black depending on the colours of the background. 

Have also corrected the emails from Valentina@agentscapital.co to Hello@agentscapital.co. 

Also focused on corrected the small heading at the bottom of the page and making them in black font with Bold writing to make them stand out.